consolidate shallow merge code
