+++ projects/gateway/pkg/translator/merge.go
@@ -3,14 +3,14 @@ package translator
 import (
 	"reflect"
 
-	"github.com/imdario/mergo"
-	"github.com/solo-io/gloo/projects/gloo/pkg/api/v1/options/hcm"
-	"github.com/solo-io/gloo/projects/gloo/pkg/api/v1/ssl"
-
 	"github.com/golang/protobuf/proto"
 	"github.com/golang/protobuf/ptypes/duration"
 	"github.com/golang/protobuf/ptypes/wrappers"
+	"github.com/imdario/mergo"
 	v1 "github.com/solo-io/gloo/projects/gloo/pkg/api/v1"
+	"github.com/solo-io/gloo/projects/gloo/pkg/api/v1/options/hcm"
+	"github.com/solo-io/gloo/projects/gloo/pkg/api/v1/ssl"
+	"github.com/solo-io/gloo/projects/gloo/pkg/utils"
 	"github.com/solo-io/solo-kit/pkg/api/v1/resources/core"
 )
 
@@ -29,7 +29,7 @@ func mergeRouteOptions(dst, src *v1.RouteOptions) *v1.RouteOptions {
 
 	for i := 0; i < dstValue.NumField(); i++ {
 		dstField, srcField := dstValue.Field(i), srcValue.Field(i)
-		shallowMerge(dstField, srcField, false)
+		utils.ShallowMerge(dstField, srcField, false)
 	}
 
 	return dst
@@ -50,49 +50,10 @@ func mergeVirtualHostOptions(dst, src *v1.VirtualHostOptions) *v1.VirtualHostOpt
 
 	for i := 0; i < dstValue.NumField(); i++ {
 		dstField, srcField := dstValue.Field(i), srcValue.Field(i)
-		shallowMerge(dstField, srcField, false)
+		utils.ShallowMerge(dstField, srcField, false)
 	}
 
 	return dst
-}
-
-// Sets dst to the value of src, if src is non-zero and dest is zero-valued or overwrite=true.
-func shallowMerge(dst, src reflect.Value, overwrite bool) {
-	if !src.IsValid() {
-		return
-	}
-
-	if dst.CanSet() && !isEmptyValue(src) && (overwrite || isEmptyValue(dst)) {
-		dst.Set(src)
-	}
-
-	return
-}
-
-// From src/pkg/encoding/json/encode.go.
-func isEmptyValue(v reflect.Value) bool {
-	switch v.Kind() {
-	case reflect.Array, reflect.Map, reflect.Slice, reflect.String:
-		return v.Len() == 0
-	case reflect.Bool:
-		return !v.Bool()
-	case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-		return v.Int() == 0
-	case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64, reflect.Uintptr:
-		return v.Uint() == 0
-	case reflect.Float32, reflect.Float64:
-		return v.Float() == 0
-	case reflect.Interface, reflect.Ptr:
-		if v.IsNil() {
-			return true
-		}
-		return isEmptyValue(v.Elem())
-	case reflect.Func:
-		return v.IsNil()
-	case reflect.Invalid:
-		return true
-	}
-	return false
 }
 
 func mergeSslConfig(parent, child *ssl.SslConfig, preventChildOverrides bool) *ssl.SslConfig {
